+++ doc/ref/examples/src/com/tightdb/refdoc/DynTableViewIntro.java
@@ -15,7 +15,7 @@ public class DynTableViewIntro {
         Table table = new Table();
 
         // Specify the column types and names
-        table.addColumn(ColumnType.LONG, "ID");
+        table.addColumn(ColumnType.INTEGER, "ID");
         table.addColumn(ColumnType.STRING, "City");
 
         // Add data to the table
