+++ doc/ref/examples/src/com/tightdb/refdoc/GroupIntro.java
@@ -18,7 +18,7 @@ public class GroupIntro {
 
         // Create a new table with 2 columns and add 3 rows of data
         Table table = group.getTable("table1");
-        table.addColumn(ColumnType.LONG, "ID");
+        table.addColumn(ColumnType.INTEGER, "ID");
         table.addColumn(ColumnType.STRING, "Animal");
         table.add(1, "Lion");
         table.add(2, "Monkey");
